refactor: move info modals to same directory
