+++ dev-client/src/components/modals/infoModals/PrivacyInfoContent.tsx
