+++ dev-client/src/screens/CreateSiteScreen/CreateSiteScreen.tsx
@@ -29,7 +29,7 @@ import {SiteAddMutationInput} from 'terraso-client-shared/graphqlSchema/graphql'
 import {ScreenScaffold} from 'terraso-mobile-client/screens/ScreenScaffold';
 import {ScreenCloseButton} from 'terraso-mobile-client/navigation/components/ScreenCloseButton';
 import {AppBar} from 'terraso-mobile-client/navigation/components/AppBar';
-import {PrivacyInfoModal} from 'terraso-mobile-client/components/infoModals/PrivacyInfoModal';
+import {PrivacyInfoModal} from 'terraso-mobile-client/components/modals/infoModals/PrivacyInfoModal';
 
 type Props =
   | {
