+++ dev-client/src/screens/ProjectViewScreen.tsx
@@ -21,7 +21,7 @@ import {ProjectTabNavigator} from 'terraso-mobile-client/navigation/navigators/P
 import {ScreenScaffold} from 'terraso-mobile-client/screens/ScreenScaffold';
 import {AppBar} from 'terraso-mobile-client/navigation/components/AppBar';
 import {useSelector} from 'terraso-mobile-client/store';
-import {PrivacyInfoModal} from 'terraso-mobile-client/components/infoModals/PrivacyInfoModal';
+import {PrivacyInfoModal} from 'terraso-mobile-client/components/modals/infoModals/PrivacyInfoModal';
 import {BottomSheetPrivacyModalContext} from 'terraso-mobile-client/context/BottomSheetPrivacyModalContext';
 import {ProjectRoleContextProvider} from 'terraso-mobile-client/context/ProjectRoleContext';
